Represent z buffer as float array

+++ include/software_renderer.h
@@ -19,7 +19,7 @@ typedef struct framebuffer {
 
 typedef struct render_state {
   framebuffer FrameBuffer;
-  i32* ZBuffer;
+  float* ZBuffer;
 } render_state;
 
 i32 RendererInit(u32 Width, u32 Height);
